+++ golem/gui/static/js/test_case_code.js
@@ -5,7 +5,7 @@ var unsavedChanges = false;
 
 $(document).ready(function() {
 
-    if(codeError.length > 0){
+    if(codeError != null){
         $(".error-container").show();
         $(".error-container pre").html(codeError);
     }
@@ -62,9 +62,9 @@ function saveTestCase(){
             unsavedChanges = false;
             codeEditor.markClean();
             utils.toast('success', "Test "+testCaseName+" saved", 3000);
-            if(data.error.length > 0){
+            if(data.error != null){
                 $(".error-container").show();
-                $(".error-container pre").html(codeError);
+                $(".error-container pre").html(data.error);
                 utils.toast('info', "There are errors in the code", 3000)
             }
             else{